requirements.txt
# Generated by running `pip install metview xarray cfgrib ecmwf-opendata psutil && pip freeze --local` in a blank env
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
cfgrib==0.9.15.0
charset-normalizer==3.4.1
click==8.1.8
eccodes==2.39.1
ecmwf-opendata==0.3.14
findlibs==0.0.5
idna==3.10
metview==1.16.1
multiurl==0.3.3
numpy==2.2.2
packaging==24.2
pandas==2.2.3
psutil==6.1.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
xarray==2025.1.1
